requirements.txt
google-auth
gTTS
pygame
speechrecognition
faster-whisper
google-generativeai
edge-tts
pydub
